fixed some accel things

+++ hardware/CowTags/radioProtocol.h
@@ -40,9 +40,9 @@ struct temperatureData {
 };
 
 struct accelerationData {
-	uint8_t x;
-	uint8_t y;
-	uint8_t z;
+	uint16_t x;
+	uint16_t y;
+	uint16_t z;
 };
 
 struct heartrateData {
